coreboot: Locate table earlier than arch_cpu_init()
Currently we locate the coreboot table in arch_cpu_init(). One of the
first things that happens in initcall_run_f() initialization is
fdtdec_setup() is called which could inspect the DTB for memory
information (see mach-snapdragon's board_fdt_blob_setup() code).
Introduce an initcall specifically to locate the coreboot table early on
in the boot process so that board_fdt_blob_setup() can know that the
memory is described in the coreboot tables instead of the DTB that's
either embedded in the ELF file or provided by the earlier stage
bootloader. This will also help in the future when we want to pick the
DTB from a FIT image based on the board and revision ID bits in the
coreboot tables.

Reviewed-by: Tom Rini <trini@konsulko.com>
Signed-off-by: Stephen Boyd <swboyd@chromium.org>