mach-snapdragon: Support booting as a coreboot payload
Coreboot sets up the DRAM, CPU, and devices before booting a payload,
meaning we can parse the coreboot tables at boot for memory info when
running as a payload. Use the memory descriptors from coreboot if the
tables are present so that Snapdragon SoCs can run U-boot as a coreboot
payload.

Reviewed-by: Caleb Connolly <caleb.connolly@linaro.org>
Signed-off-by: Stephen Boyd <swboyd@chromium.org>