board: google: Add board_debug_uart_init() for Trogdor
Implement board_debug_uart_init() for Trogdor boards so that the debug
UART can be used. Initialize the clk to XO speed (19.2MHz) and change
the pin configuration to mux out the uart hardware on the pins that are
connected to the closed case debug (CCD) port for the AP. This is enough
to get the early debug console working on boards like Lazor, but it
should work for all Google Trogdor boards.

Signed-off-by: Stephen Boyd <swboyd@chromium.org>