doc: chromebook_trogdor: Add document detailing installation steps
Make it a little more ergonomic than navigating to the chromium.org docs
by just documenting the steps needed to install U-Boot on a Trogdor
chromebook. I've been using the Depthcharge 'altfw' flow to do this
which is nice because you get to keep ChromeOS installed for any future
firmware updates.

Signed-off-by: Stephen Boyd <swboyd@chromium.org>